linux build

+++ src/log.c
@@ -24,7 +24,7 @@ char* log_get_current_time() {
     tm_info = gmtime(&tv.tv_sec);
 
     snprintf(time_str, sizeof(time_str), "%02d:%02d:%02d.%03d",
-             tm_info->tm_hour, tm_info->tm_min, tm_info->tm_sec, tv.tv_usec / 1000);
+             tm_info->tm_hour, tm_info->tm_min, tm_info->tm_sec, (int)(tv.tv_usec / 1000));
 #endif
 
   return time_str;
